Move chemical reaction effects into Chemistry/ReactionEffects/ subfolder
