+++ Content.Server/Interfaces/Chemistry/IReactionEffect.cs
@@ -8,6 +8,6 @@ namespace Content.Shared.Interfaces
     /// </summary>
     public interface IReactionEffect : IExposeData
     {
-        void React(IEntity solutionEntity, int intensity );
+        void React(IEntity solutionEntity, int intensity);
     }
 }
